% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate.R
\name{Calculate}
\alias{Calculate}
\title{Calculate the correct rate of the result of selecting \eqn{d_1}, \eqn{d_2}, \eqn{d_3}}
\usage{
Calculate(true, results)
}
\arguments{
\item{true}{Vector. The true number of clusters in each mode (eg. c(3,3,3)).}

\item{results}{List. A list consists of 3*1 matrix and each matrix is the estimated c(d_1,d_2,d_3). (The return value of sparse.choosekrl().)}
}
\description{
Calculate the correct rate of the result of sparse.choosekrl().
}
