% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculatekrl.R
\name{Calculatekrl}
\alias{Calculatekrl}
\title{summary the result of sparse.choosekrl}
\usage{
Calculatekrl(results)
}
\arguments{
\item{results}{result returned by sparse.choosekrl()}
}
\value{
\code{meank}
        \code{meanr}
        \code{meanl}
        \code{sdk}
        \code{sdr}
        \code{sdl}
}
\description{
summary the result of sparse.choosekrl
}
