% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cer.R
\name{cer}
\alias{cer}
\title{Calculate the clustering error rate (CER)}
\usage{
cer(bires, data)
}
\arguments{
\item{bires}{List. The return value of label2() or classify2() which consists of judgeX, Cs, Ds, Es, objs, mus.}

\item{data}{List. The return value of get.data().}
}
\value{
A list which consists of the CER of all three modes.
}
\description{
Calculate clustering error rate (CER) by comparing the clustering result and the true underlying mean signal tensor.
}
