% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_kmeans.R
\name{cp_kmeans}
\alias{cp_kmeans}
\title{Perform tensor clustering via cp decomposition}
\usage{
cp_kmeans(x, k, r, l, multiplicative = NULL, max.s = NULL)
}
\arguments{
\item{x}{a three-dimensional array; data tensor}

\item{k}{\eqn{d_1}: the clusters number of mode 1}

\item{r}{\eqn{d_2}: the clusters number of mode 2}

\item{l}{\eqn{d_3}: the clusters number of mode 3}

\item{multiplicative}{the number of components}

\item{max.s}{the max value when selecting s}
}
\value{
a list with only one element: judgeX
}
\description{
Perform tensor clustering via cp decomposition.
}
