% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tensor.R
\name{plot_tensor}
\alias{plot_tensor}
\title{Draw 3d plot of tensor}
\usage{
plot_tensor(tensor)
}
\arguments{
\item{tensor}{a three-dimensional array}
}
\value{
None
}
\description{
Draw 3d plot of a given tensor.
}
\examples{
x = get.data(30,30,30,3,3,3)$x
plot_tensor(x)
}
