% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.choosekrl.R
\name{sim.choosekrl}
\alias{sim.choosekrl}
\title{Perform simulation: selecting the best \eqn{d_1}, \eqn{d_2} and \eqn{d_3}}
\usage{
sim.choosekrl(n, p, q, k, r, l, error = 1, sim.times = 5,
  method = "L0", mode = "bic", seed = TRUE)
}
\arguments{
\item{n}{the dimension of mode 1}

\item{p}{the dimension of mode 2}

\item{q}{the dimension of mode 3}

\item{k}{\eqn{d_1}: the clusters number of mode 1}

\item{r}{\eqn{d_2}: the clusters number of mode 2}

\item{l}{\eqn{d_3}: the clusters number of mode 3}

\item{error}{positive numeric value. The noise when producing the data}

\item{sim.times}{positive integer. Simulation times}

\item{method}{two options: "L0", "L1". Two methods use different penalties, where "L0" indicating L0 penalty and "L1" indicating Lasso penalty.}

\item{mode}{two options: "bic": selecting cluster numbers by bic; "crossvalidation":selecting cluster numbers by cross validation.}

\item{seed}{logic value. Whether set seed to each simulation.}
}
\value{
A list consists of the estimating result in different iteration.
}
\description{
Simulation: selecting the best \eqn{d_1}, \eqn{d_2} and \eqn{d_3}.
}
