% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.choosekrl.R
\name{sparse.choosekrl}
\alias{sparse.choosekrl}
\title{choose the best k,r,l to do clustering when the ranges are given}
\usage{
sparse.choosekrl(x, k, r, l, lambda = 0, percent = 0.2,
  trace = FALSE, nstart = 20, sim.times = 1, method = "L0")
}
\arguments{
\item{x}{a three-dimensional array}

\item{k}{a vector, the possible clusters number of mode 1}

\item{r}{a vector, the possible clusters number of mode 2}

\item{l}{a vector, the possible clusters number of mode 3}

\item{lambda}{a numeric value}

\item{percent}{a numeric value between 0 and 1}

\item{trace}{...}

\item{nstart}{...}

\item{sim.times}{the same as label2()}

\item{method}{two options: "L0", "L1". Two methods use different penalties, where "L1" indicating Lasso penalty.}
}
\value{
a list \code{estimated_krl} ...\\
               \code{results.se} ...\\
               \code{results.maen} ...\\
}
\description{
choose the best k,r,l to do clustering when the ranges are given using cross validation.
}
