% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.evaluate.R
\name{sparse.evaluate}
\alias{sparse.evaluate}
\title{evaluate the accuracy of the model when inputting a sparse tensor}
\usage{
sparse.evaluate(bires, data, CER = TRUE, show = TRUE)
}
\arguments{
\item{bires}{the return list of label2()}

\item{data}{the return list of get.data()}

\item{CER}{if true, it would return CER also}

\item{show}{whether print the result}
}
\description{
evaluate the accuracy of the model when inputting a sparse tensor
}
