% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tucker_kmeans.R
\name{tucker_means}
\alias{tucker_means}
\title{Tensor clustering by performing k-means on tucker decomposition}
\usage{
tucker_means(x, k, r, l)
}
\arguments{
\item{x}{a three-dimensional array}

\item{k}{numeric value. The clusters number of mode 1}

\item{r}{numeric value. The clusters number of mode 2}

\item{l}{numeric value. The clusters number of mode 3}
}
\value{
a list with only one element: judgeX
}
\description{
Tensor clustering by performing k-means on tucker decomposition.
}
