% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculatekrl.R
\name{Calculatekrl}
\alias{Calculatekrl}
\title{Summarize the result of estimating the true c(d_1,d_2,d_3) in a simulation.}
\usage{
Calculatekrl(results)
}
\arguments{
\item{results}{List. A list consists of 3-dimensional vectors and each vector is estimated c(d_1,d_2,d_3). (The result returned by sparse.choosekrl().)}
}
\value{
\code{meank} mean estimated d_1
        \code{meanr} mean estimated d_2
        \code{meanl} mean estimated d_3
        \code{sdk} the standard deviation of estimated d_1
        \code{sdr} the standard deviation of estimated d_2
        \code{sdl} the standard deviation of estimated d_3
}
\description{
Summarize the return objct of sparse.choosekrl().
}
