% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choosekrl_bic.R
\name{choosekrl_bic}
\alias{choosekrl_bic}
\title{choose the best k,r,l to do clustering when the ranges are given}
\usage{
choosekrl_bic(x, k, r, l, lambda = 0, sim.times = 1, method = "L0")
}
\arguments{
\item{x}{a three-dimensional array}

\item{k}{a vector, the possible clusters number of mode 1}

\item{r}{a vector, the possible clusters number of mode 2}

\item{l}{a vector, the possible clusters number of mode 3}

\item{lambda}{a numeric value}

\item{sim.times}{the same as label2()}

\item{method}{two options: "L0", "L1". Two methods use different penalties, where "L1" indicating Lasso penalty.}
}
\value{
a list \code{estimated_krl} ...\\
               \code{BIC} ...
}
\description{
choose the best k,r,l to do clustering when the ranges are given by minimizing the BIC.
}
