% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{Perform simulation: calculating the mean CER over iterations}
\usage{
simulation(n, p, q, k, r, l, error, lambda, iteration = 1,
  method = "L0")
}
\arguments{
\item{n}{the dimension of mode 1}

\item{p}{the dimension of mode 2}

\item{q}{the dimension of mode 3}

\item{k}{\eqn{d_1}: the clusters number of mode 1}

\item{r}{\eqn{d_2}: the clusters number of mode 2}

\item{l}{\eqn{d_3}: the clusters number of mode 3}

\item{error}{the standard deviation when producing data}

\item{lambda}{...}

\item{iteration}{iteration times}

\item{method}{two options: "L0", "L1". Two methods use different penalties, where "L1" indicating Lasso penalty.}
}
\description{
Perform simulation: calculating the mean CER over iterations.
}
