% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.evaluate.R
\name{sparse.evaluate}
\alias{sparse.evaluate}
\title{Evaluate the accuracy of clustering result}
\usage{
sparse.evaluate(bires, data, CER = TRUE, show = TRUE)
}
\arguments{
\item{bires}{the return value of label2()}

\item{data}{the return value of get.data()}

\item{CER}{logic value. If true, it would return CER also}

\item{show}{logic value. If true, it would print the result}
}
\description{
Given the input tensor, perform tensor clustering and evaluate the accuracy of the clustering result.
}
